Customer-Service_Account&Card_Service_done

+++ Customer-Service/src/main/java/de/grado/customerservice/repository/CardsRepository.java
@@ -4,12 +4,13 @@ import de.grado.customerservice.model.Cards;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CardsRepository extends JpaRepository<Cards, Long>
+public interface CardsRepository extends JpaRepository<Cards, BigInteger>
 {
-    List<Cards> findAllByAccount_Id(Long accountId);
+    List<Cards> findAllByAccount_Id(BigInteger accountId);
     Optional<Cards> findByCardNumber(String cardNumber);
 }
